refactor: move listprocessesWithMemoryUsageUnix file to listprocessesWithMemoryUsageUnix folder
