+++ packages/main-process/src/parts/ListProcessesWithMemoryUsage/ListProcessesWithMemoryUsage.js
@@ -2,9 +2,9 @@ const Platform = require('../Platform/Platform.js')
 
 const getModule = () => {
   if (Platform.isWindows) {
-    return require('./ListProcessesWithMemoryUsageWindows.js')
+    return require('../ListProcessesWithMemoryUsageWindows/ListProcessesWithMemoryUsageWindows.js')
   }
-  return require('./ListProcessesWithMemoryUsageUnix.js')
+  return require('../ListProcessesWithMemoryUsageUnix/ListProcessesWithMemoryUsageUnix.js')
 }
 
 exports.listProcessesWithMemoryUsage = async (rootPid) => {
